move videoframechecker to separate file and export as module
moves the videoframechecker to a separate file and allows requiring
videoframechecker and ssim as modules. Also adds an index.js for module usage.
