+++ src/index.html
@@ -51,6 +51,7 @@
   <script src="js/call.js"></script>
   <script src="js/stats.js"></script>
   <script src="js/ssim.js"></script>
+  <script src="js/videoframechecker.js"></script>
   <script src="js/util.js"></script>
   <script src="js/testcasename.js"></script>
   <script src="js/testsuitename.js"></script>
