+++ src/js/camresolutionstest.js
@@ -289,79 +289,3 @@ CamResolutionsTest.prototype = {
   }
 };
 
-// TODO: Move this to a separate file.
-function VideoFrameChecker(videoElement) {
-  this.frameStats = {
-    numFrozenFrames: 0,
-    numBlackFrames: 0,
-    numFrames: 0
-  };
-
-  this.running_ = true;
-
-  this.nonBlackPixelLumaThreshold = 20;
-  this.previousFrame_ = [];
-  this.identicalFrameSsimThreshold = 0.985;
-  this.frameComparator = new Ssim();
-
-  this.canvas_ = document.createElement('canvas');
-  this.videoElement_ = videoElement;
-  this.listener_ = this.checkVideoFrame_.bind(this);
-  this.videoElement_.addEventListener('play', this.listener_, false);
-}
-
-VideoFrameChecker.prototype = {
-  stop: function() {
-    this.videoElement_.removeEventListener('play' , this.listener_);
-    this.running_ = false;
-  },
-
-  getCurrentImageData_: function() {
-    this.canvas_.width = this.videoElement_.width;
-    this.canvas_.height = this.videoElement_.height;
-
-    var context = this.canvas_.getContext('2d');
-    context.drawImage(this.videoElement_, 0, 0, this.canvas_.width,
-        this.canvas_.height);
-    return context.getImageData(0, 0, this.canvas_.width, this.canvas_.height);
-  },
-
-  checkVideoFrame_: function() {
-    if (!this.running_) {
-      return;
-    }
-    if (this.videoElement_.ended) {
-      return;
-    }
-
-    var imageData = this.getCurrentImageData_();
-
-    if (this.isBlackFrame_(imageData.data, imageData.data.length)) {
-      this.frameStats.numBlackFrames++;
-    }
-
-    if (this.frameComparator.calculate(this.previousFrame_, imageData.data) >
-        this.identicalFrameSsimThreshold) {
-      this.frameStats.numFrozenFrames++;
-    }
-    this.previousFrame_ = imageData.data;
-
-    this.frameStats.numFrames++;
-    setTimeout(this.checkVideoFrame_.bind(this), 20);
-  },
-
-  isBlackFrame_: function(data, length) {
-    // TODO: Use a statistical, histogram-based detection.
-    var thresh = this.nonBlackPixelLumaThreshold;
-    var accuLuma = 0;
-    for (var i = 4; i < length; i += 4) {
-      // Use Luma as in Rec. 709: Y′709 = 0.21R + 0.72G + 0.07B;
-      accuLuma += 0.21 * data[i] + 0.72 * data[i + 1] + 0.07 * data[i + 2];
-      // Early termination if the average Luma so far is bright enough.
-      if (accuLuma > (thresh * i / 4)) {
-        return false;
-      }
-    }
-    return true;
-  }
-};
